Split ChangePassword into components
